% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{analyze_gene_signatures}
\alias{analyze_gene_signatures}
\title{Calculate relationships between cell types}
\usage{
analyze_gene_signatures(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  cell.types.use = NULL,
  return.objs = F
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{cell.types.use}{A string of cell type labels to include in the plot,
by default all cell types present}

\item{return.objs}{Logical, whether to return a list of matrices of
derived data}
}
\value{
named list of cosine similarity matrix and hierarchical clustering,
if \code{return.objs = TRUE}
}
\description{
Calculate relationships between cell types
}
