% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{analyze_spatial_correlation}
\alias{analyze_spatial_correlation}
\title{Calculate relationships between cell type distributions}
\usage{
analyze_spatial_correlation(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  mat.use = "proportions",
  cell.types.use = NULL,
  return.objs = F
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{mat.use}{A string, either "raw" or "proportions"
referring to what version of the results to summarize}

\item{cell.types.use}{A string of cell type labels to include in the plot,
by default all cell types present}

\item{return.objs}{Logical, whether to return a list of matrices of
derived data}
}
\value{
named list of pearson correlation matrix and hierarchical clustering,
if \code{return.objs = TRUE}
}
\description{
Calculate relationships between cell type distributions
}
