% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_processing.R
\name{calculate_cell_sizes}
\alias{calculate_cell_sizes}
\title{Calculate cell sizes with all reference data}
\usage{
calculate_cell_sizes(objs, annotations, filepath, analysis.name, plot.hist = F)
}
\arguments{
\item{objs}{Either a list of matrices,
a list of paths to RDS files,
a list of paths to H5 files,
or liger objects}

\item{annotations}{Named vector of cell type assignments by sample}

\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{plot.hist}{Logical, if to display and save histograms
of nUMIs by cell type}

\item{n.cells}{Integer value corresponding to maximum number
of samples per cell type}

\item{rand.seed}{Integer random seed}
}
\value{
nothing
}
\description{
Calculate cell sizes with all reference data
}
