% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{calculate_wasserstein}
\alias{calculate_wasserstein}
\title{Calculate the Wasserstein distance between cell-types and genes}
\usage{
calculate_wasserstein(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  mat.use = "proportions",
  cell.types.use = NULL,
  genes.use = NULL,
  p = 2,
  min.samples = 1,
  return.objs = F
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{mat.use}{A string, either "raw" or "proportions"
referring to what version of the results to summarize}

\item{cell.types.use}{A string of cell type labels to include in the plot,
by default all cell types present}

\item{genes.use}{A string of genes to include in a plot, by default none}

\item{p}{The \code{p} exponent used for the Minkowski distance}

\item{min.samples}{Integer value, the minimum number of samples a cell type
can load on and be included in the analysis}

\item{return.objs}{Logical, whether to return a list of matrices of
derived data}
}
\value{
matrix of pairwise Wasserstein distances if \code{return.objs = TRUE}
}
\description{
Calculate the Wasserstein distance between cell-types and genes
}
