% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{cell_type_loading_histogram}
\alias{cell_type_loading_histogram}
\title{Generate histograms of loading by cell type}
\usage{
cell_type_loading_histogram(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  mat.use = "proportions",
  cell.types.plot = NULL,
  print.plots = TRUE,
  bin.num = 30
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{mat.use}{A string, either "raw" or "proportions"
referring to what version of the results to summarize}

\item{print.plots}{Logical, whether to display results in the plots panel}

\item{bin.num}{Integer number of bins to use in histogram}

\item{cell.types.use}{A string of cell type labels to include in the plot,
by default all cell types present}
}
\value{
nothing
}
\description{
Generate histograms of loading by cell type
}
