% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolve_spatial}
\alias{deconvolve_spatial}
\title{Title}
\usage{
deconvolve_spatial(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  cell.size = T
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{cell.size}{Logical, if to scale gene signatures by cell sizes}
}
\value{
nothing
}
\description{
Title
}
