% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plotting.R
\name{generate_label_gifs}
\alias{generate_label_gifs}
\title{Title}
\usage{
generate_label_gifs(
  filepath,
  analysis.name,
  spatial.data.name,
  labels.plot,
  dims = c(500, 500)
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{labels.plot}{A named vector or matrix of labels to plot for the
provided coordinates}

\item{dims}{Integer vector of length 2 corresponding to the width and
height of the RGL window}
}
\description{
Title
}
