% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{learn_gene_signatures}
\alias{learn_gene_signatures}
\title{Title}
\usage{
learn_gene_signatures(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  lambda = 1,
  thresh = 1e-08,
  max.iters = 100,
  nrep = 1,
  print.obj = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{lambda}{Double, regularization parameter for which increasing penalizes
dataset-specific effects}

\item{thresh}{Double, minimum fractional change in objective function to
continue iteration}

\item{max.iters}{Integer maximum of iterations to complete before pausing}

\item{nrep}{Number of random starts to complete}

\item{print.obj}{Logical, if to print current value of objective}

\item{verbose}{Logical, if to print the final objective and best random seed}
}
\value{
nothing
}
\description{
Title
}
