% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_processing.R
\name{load_objs}
\alias{load_objs}
\title{Load data from one of multiple formats}
\usage{
load_objs(objs, datasets.remove)
}
\arguments{
\item{objs}{A named list of matrices (dgCMatrix), RDS file paths to matirces,
H5 file paths to LIGER analyzed datasets.}
}
\value{
list object. List element type depends on input.
}
\description{
Load data from one of multiple formats
}
