% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{mirror_spatial_coords}
\alias{mirror_spatial_coords}
\title{Flip axes in spatial data}
\usage{
mirror_spatial_coords(
  filepath,
  analysis.name,
  spatial.data.name,
  axes.flip = c(FALSE, FALSE, FALSE),
  overwrite = T
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{axes.flip}{A vector with three logicals, corresponding to which of
the axes to invert}

\item{overwrite}{Logical, if the original data should be overwritten,
otherwise "\code{spatial.data.name}_mirror\verb{_x}/\verb{_y},\verb{_z}is created}
}
\value{
nothing
}
\description{
Flip axes in spatial data
}
