% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plotting.R
\name{overlay_subregion_gifs}
\alias{overlay_subregion_gifs}
\title{Title}
\usage{
overlay_subregion_gifs(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  mat.use = "proportions",
  cell.types.plot = NULL,
  subregions.plot = NULL,
  filter = NULL,
  dims = c(500, 500)
)
}
\arguments{
\item{filepath}{filepath}

\item{analysis.name}{analysis.name}

\item{spatial.data.name}{spatial.data.name}

\item{rand.seed}{rand.seed}

\item{mat.use}{mat.use}

\item{cell.types.plot}{cell.types.plot}

\item{subregions.plot}{subregions.plot}

\item{filter}{filter}

\item{dims}{dims}
}
\value{
nothing
}
\description{
Title
}
