% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{plot_calculate_wasserstein}
\alias{plot_calculate_wasserstein}
\title{Plot results of \code{calculate_wasserstein}}
\usage{
plot_calculate_wasserstein(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  print.plots = T
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{print.plots}{Logical, whether to display results in the plots panel}
}
\value{
nothing
}
\description{
Plot results of \code{calculate_wasserstein}
}
