% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{qc_spatial_data}
\alias{qc_spatial_data}
\title{Quality-control spatial data}
\usage{
qc_spatial_data(
  filepath,
  analysis.name,
  spatial.data.name,
  count.data = FALSE,
  z = 1,
  n.umi.thresh = 150,
  rand.seed = 123
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{count.data}{Logical, if the spatial data is from a counts
or intensity-based modality}

\item{z}{Double, the standard deviations above the mean that the number of
NAs in a gene can be before the gene is removed, for intensity data}

\item{n.umi.thresh}{Integer number of counts below which to remove a sample,
for counts based data}

\item{rand.seed}{Integer random seed}
}
\value{
nothing
}
\description{
Quality-control spatial data
}
