% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{reference_3d_coordinates}
\alias{reference_3d_coordinates}
\title{Generate silouhettes of the data along all three axes}
\usage{
reference_3d_coordinates(
  filepath,
  analysis.name,
  spatial.data.name,
  save.plots = FALSE
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{save.plots}{A logical, corresponding with if to save requested plots upon
generation}
}
\value{
nothing
}
\description{
Generate silouhettes of the data along all three axes
}
\examples{
\dontrun{

}
}
