% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{register_voxel_to_label}
\alias{register_voxel_to_label}
\title{Transfer labels from coarse-grained sampled}
\usage{
register_voxel_to_label(
  filepath,
  analysis.name,
  spatial.data.name,
  labels.use,
  label.name
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{labels.use}{Named vector of labels for the prevoxelized data}

\item{label.name}{String identifying the label set}
}
\value{
nothing
}
\description{
Transfer labels from coarse-grained sampled
}
