% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{save_spatial_data}
\alias{save_spatial_data}
\title{Add a new spatial dataset to the analysis directory}
\usage{
save_spatial_data(
  filepath,
  analysis.name,
  spatial.data.file,
  coords.file,
  spatial.data.name
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.file}{Path to an RDS file containing desired expression data}

\item{coords.file}{Path to an RDS file containing desired coordinate data}

\item{spatial.data.name}{String identifying the spatial sample}
}
\value{
nothing
}
\description{
Add a new spatial dataset to the analysis directory
}
