% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_processing.R
\name{select_defining_genes}
\alias{select_defining_genes}
\title{select variable genes with the Kruskal-Wallis test}
\usage{
select_defining_genes(
  filepath,
  analysis.name,
  deconv.gene.num = 2000,
  gene.num.tol = 50,
  rand.seed = 123
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{deconv.gene.num}{Integer, the number of genes to select}

\item{gene.num.tol}{Integer, the maximum difference between the number of
genes selected and \code{deconv.gene.num}}

\item{rand.seed}{Integer random seed}
}
\value{
nothing
}
\description{
select variable genes with the Kruskal-Wallis test
}
