% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiNMFiD-package.R
\docType{package}
\name{SiNMFiD-package}
\alias{SiNMFiD}
\alias{SiNMFiD-package}
\title{SiNMFiD: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joshua Sodicoff \email{sodicoff@uchicago.edu}

}
\keyword{internal}
