% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{start_analysis}
\alias{start_analysis}
\title{Set up new analysis directory}
\usage{
start_analysis(filepath, analysis.name)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}
}
\value{
nothing
}
\description{
Set up new analysis directory
}
