% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{subset_spatial_data}
\alias{subset_spatial_data}
\title{Subset a spatial dataset by coordinates for analysis}
\usage{
subset_spatial_data(
  filepath,
  analysis.name,
  spatial.data.name,
  subset.specs = list(c(NaN, NaN), c(NaN, NaN), c(NaN, NaN)),
  new.spatial.data.name = NULL,
  out.filepath = NULL
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{subset.specs}{A list with length equal to the number of axes, with
each entry a vector of length two, with the first element being the
minimum value to include and the second being the maximum, or NaN to
indicate a missing value}

\item{new.spatial.data.name}{String, optional name for new analysis, otherwise
the default "\code{spatial.data.name}\emph{subset}\code{n.samples}" is used}

\item{out.filepath}{Path to directory to save subset data to
if not within the analysis}
}
\value{
nothing
}
\description{
Subset a spatial dataset by coordinates for analysis
}
