% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{summarize_by_layer}
\alias{summarize_by_layer}
\title{Summarize cell-type and gene expression data by}
\usage{
summarize_by_layer(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  layer.list,
  type = "mean",
  mat.use = "proportions",
  cell.types.use = NULL,
  genes.use = NULL,
  return.objs = FALSE
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{layer.list}{A named list of spatial samples by layer of interest}

\item{type}{A string, either "mean" or "sum", how results should be
combined for summary}

\item{mat.use}{A string, either "raw", "proportions", or "assignments"
referring to what version of the results to summarize}

\item{cell.types.use}{A string of cell type labels to include in the plot,
by default all cell types present}

\item{genes.use}{A string of genes to include in a plot, by default none}

\item{return.objs}{Logical, whether to return a list of matrices of
derived data}
}
\value{
cell-type and gene expression data summarized by layer in a named
list, if \code{return.objs = TRUE}
}
\description{
Summarize cell-type and gene expression data by
}
