% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{summarize_clusters}
\alias{summarize_clusters}
\title{Summarize cell types present in the source annotations}
\usage{
summarize_clusters(filepath, analysis.name, return.objs = F)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{return.objs}{Logical, whether to return a vector of the names of clusters}
}
\value{
A vector of unique clusters in the source annotations,
if \code{return.objs = TRUE}
}
\description{
Summarize cell types present in the source annotations
}
