% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{summarize_subregions}
\alias{summarize_subregions}
\title{Summarize subregions of a vector of regions of interest}
\usage{
summarize_subregions(
  regions,
  ontology.file = "Downloads/allen_structure_ontology.csv",
  return.objs = F
)
}
\arguments{
\item{regions}{A vector of region names}

\item{ontology.file}{A csv describing the Allen structure ontology}

\item{return.objs}{Logical, whether to return acronyms for all subregions found}
}
\value{
A vector of unique subregions within the provided regions,
if \code{return.objs = TRUE}
}
\description{
Summarize subregions of a vector of regions of interest
}
