% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plotting.R
\name{three_d_plotting_routine}
\alias{three_d_plotting_routine}
\title{Routine for loading filepaths relevant to 3d plotting functionality}
\usage{
three_d_plotting_routine(filepath, analysis.name, spatial.data.name, rand.seed)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}
}
\description{
Routine for loading filepaths relevant to 3d plotting functionality
}
