% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plotting.R
\name{three_d_plotting_routine_end}
\alias{three_d_plotting_routine_end}
\title{Clear relevant paths out of environment}
\usage{
three_d_plotting_routine_end()
}
\value{
nothing
}
\description{
Clear relevant paths out of environment
}
