% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{transform_coords_to_ccf}
\alias{transform_coords_to_ccf}
\title{Use predefined transformations to match some modalities to the Allen CCF}
\usage{
transform_coords_to_ccf(filepath, analysis.name, spatial.data.name, ish = T)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{ish}{Logical, if the data comes from the Allen Institute
quantified ISH dataset}
}
\value{
nothing
}
\description{
Use predefined transformations to match some modalities to the Allen CCF
}
