% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plotting.R
\name{view_in_rgl}
\alias{view_in_rgl}
\title{Title}
\usage{
view_in_rgl(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  cell.type,
  mat.use = "proportions",
  filter.samples = NULL,
  dims = c(500, 500)
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{rand.seed}{Integer random seed}

\item{cell.type}{A string corresponding to one cell type found in
the deconvolution results}

\item{mat.use}{A string, either "raw" or "proportions"
referring to what version of the results to summarize}

\item{filter.samples}{Value for binarizing results, either presence above
the provided threshold or absence below}

\item{dims}{Integer vector of length 2 corresponding to the width and
height of the RGL window}
}
\value{
nothing
}
\description{
Title
}
