% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_preprocessing.R
\name{voxelize_single_cells}
\alias{voxelize_single_cells}
\title{Coarse-grain spatial data to a predetermined resolution}
\usage{
voxelize_single_cells(
  filepath,
  analysis.name,
  spatial.data.name,
  voxel.size,
  out.filepath = NULL,
  verbose = TRUE
)
}
\arguments{
\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{spatial.data.name}{String identifying the spatial sample}

\item{voxel.size}{Integer, side length of one voxel}

\item{out.filepath}{Path to directory to save subset data to
if not within the analysis}

\item{verbose}{Logical, if to print several lines of metadata on results}
}
\value{
nothing
}
\description{
Coarse-grain spatial data to a predetermined resolution
}
