% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_processing.R
\name{load_objs}
\alias{load_objs}
\title{Load data from one of multiple formats}
\usage{
load_objs(objs)
}
\arguments{
\item{objs}{Either a list of matrices,
a list of paths to RDS files,
a list of paths to H5 files,
or liger objects}
}
\value{
nothing
}
\description{
Load data from one of multiple formats
}
