% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_plotting.R
\name{overlay_subregion_gifs}
\alias{overlay_subregion_gifs}
\title{Title}
\usage{
overlay_subregion_gifs(
  filepath,
  analysis.name,
  spatial.data.name,
  rand.seed = 123,
  mat.use = "proportions",
  cell.types.plot = NULL,
  subregions.plot = NULL,
  filter = NULL,
  dims = c(500, 500)
)
}
\arguments{
\item{dims}{}
}
\description{
Title
}
