% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_processing.R
\name{sample_single_cell}
\alias{sample_single_cell}
\title{Sample from single cell reference datasets}
\usage{
sample_single_cell(
  objs,
  annotations,
  filepath,
  analysis.name,
  datasets.remove = NULL,
  n.cells = 500,
  rand.seed = 123
)
}
\arguments{
\item{objs}{Either a list of matrices,
a list of paths to RDS files,
or liger objects}

\item{annotations}{Named vector of cell type assignments by sample}

\item{filepath}{Path to analysis directory}

\item{analysis.name}{String identifying the analysis}

\item{datasets.remove, }{Named list of datasets to remove, if a list of
online liger objects is provided}

\item{n.cells}{Integer value corresponding to maximum number
of samples per cell type}

\item{rand.seed}{Integer random seed}
}
\value{
nothing
}
\description{
Sample from single cell reference datasets
}
